added a command and a query

+++ EmployeeCleanArch.Application/Interfaces/Repositories/IEmployeeRepository.cs
@@ -4,6 +4,6 @@ namespace EmployeeCleanArch.Application.Interfaces.Repositories
 {
     public interface IEmployeeRepository
     {
-        Task<List<Employee>> GetEmployeesByDepartmentAsync(int departmentId);
+        Task<List<Employee>> GetEmployeesByDepartmentAsync(long departmentId);
     }
 }
